Add file Recorded+Event.swift
